requirements.txt
streamlit==1.5.0
tensorflow==2.5.0
opencv-python-headless==4.5.2.54
numpy==1.23.4
Pillow==9.0.0
